Fix project_cost_report tests
